Get particles outside of Sampler class
This will be necessary to apply the same sampler to different patches.
Sampler should be constructed once, do all the preparations (saving
integrals, arranging all triplets/pairs with given quantum numbers, etc)
and then it should be called for different patches with different particles.

So far the functionality did not change. This commit is only structural.

Signed-off-by: Dmytro Oliinychenko <doliinychenko@lbl.gov>