Make counters atomic for parallel for
It wasn't affecting sampling, but counters worked incorretly
in case of many threads. Now it's fixed.

I have also learned that cout << a << b << endl can be
shared between threads, while printf('%d %d', a, b) is a single
operation. So I have replace cout to printf in the places, where
for many threads the code was printing nonsense.

Signed-off-by: Dmytro Oliinychenko <doliinychenko@lbl.gov>